DataType changed

+++ BudgetPlan.Shared/Dtos/BudgetPlanDetailsDto.cs
@@ -2,8 +2,8 @@ namespace BudgetPlan.Shared.Dtos;
 
 public class BudgetPlanDetailsDto
 {
-    public float AmountAllocated { get; set; } = 0f;
-    public float ValueOfTransactionItemsForDays =>
+    public double AmountAllocated { get; set; } = 0f;
+    public double ValueOfTransactionItemsForDays =>
         TransactionItemsForDaysDtos.Sum(x => x.ValuesOfTransactionItemsForDay);
     public List<TransactionItemsForDayDto> TransactionItemsForDaysDtos { get; set; } = new();
 }
